Move away from api helpers to query helpers
