+++ packages/discovery-provider/src/api/v1/authorization_helpers.py
